IMO, the words "makes us better" is more consistent (with innovation's position on the list), is the truth, and describes a purpose.
The statement 'we innovate therefore we are' is inconsistent with it's ordinal position on the list.

Our main thing is people. #1
Why we're here is excellence at delivery. #2

Innovation is #5

Not to say it isn't important, that's why it's on the list. But it isn't worthy of the title "therefore we are".
If we're going to use that parlance we should have:

We develop people therefore we are
We deliver therefore we are
...